third_party/blink/web_tests/external/wpt/cookies/value/value-ctl-expected.txt

This is a testharness.js-based test.
Found 4 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Cookie with %x9 in value is accepted (DOM).
  assert_equals: The cookie was set as expected. expected "test=9\\tvalue" but got ""
[FAIL] Cookie with %x9 in value is accepted (HTTP).
  assert_equals: The cookie was set as expected. expected "test=9\\tvalue" but got ""
[FAIL] Cookie with %xa in name is rejected or modified (HTTP).
  assert_equals: The cookie was set as expected. expected "test10 name=10" but got "test10"
[FAIL] Cookie with %xd in name is rejected or modified (HTTP).
  assert_equals: The cookie was set as expected. expected "test13 name=13" but got "test13"
Harness: the test ran to completion.
